+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] - 2020-03-10
+
+### Fixed
+- Attribute list parsing
+
+### Changed
+- Warn on fixup mismatch instead of fail
+
+### Added
+- Additional File Attribute flags
+- `valid_fixup` field to MftEntry
+
 ## [0.5.1] - 2020-02-06
 
 ### Fixed